Rename cli/cli.ts to cli/index.ts
