+++ src/index.ts
@@ -1,4 +1,4 @@
 export { createProgram, Program, Reporting } from './program'
 export { Diagnostic, Severity } from './diagnostic'
 export { Binding, BindingData, BindingExpression, BindingName, DiagNode, Document, Node, NodeType, ImportNode } from './parser/syntax-tree'
-export { ConfigOptions } from './cli/cli'
+export { ConfigOptions } from './cli'
